+++ SprintPlanner.ConsoleExample/Program.cs
@@ -1,9 +1,10 @@
 ﻿using SprintPlanner.Core;
 using System;
+using System.Security;
 
 namespace SprintPlanner
 {
-    public class Program
+    public static class Program
     {
         public static void Main(string[] args)
         {
@@ -17,7 +18,7 @@ namespace SprintPlanner
                 string username = Console.ReadLine();
 
                 Console.Write("Please enter password: ");
-                string password = Console.ReadLine();
+                var password = GetPassword();
 
                 isLoggedIn = jh.Login(username, password);
                 Console.WriteLine("--------------------");
@@ -53,6 +54,33 @@ namespace SprintPlanner
 
 
             Console.ReadLine();
+        }
+
+        public static SecureString GetPassword()
+        {
+            var pwd = new SecureString();
+            while (true)
+            {
+                ConsoleKeyInfo i = Console.ReadKey(true);
+                if (i.Key == ConsoleKey.Enter)
+                {
+                    break;
+                }
+                else if (i.Key == ConsoleKey.Backspace)
+                {
+                    if (pwd.Length > 0)
+                    {
+                        pwd.RemoveAt(pwd.Length - 1);
+                        Console.Write("\b \b");
+                    }
+                }
+                else if (i.KeyChar != '\u0000') // KeyChar == '\u0000' if the key pressed does not correspond to a printable character, e.g. F1, Pause-Break, etc
+                {
+                    pwd.AppendChar(i.KeyChar);
+                    Console.Write("*");
+                }
+            }
+            return pwd;
         }
     }
 }