+++ SprintPlanner.Core/Extensions/SecureStringExtensions.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using System.Security;
+
+namespace SprintPlanner.Core.Extensions
+{
+    public static class SecureStringExtensions
+    {
+        public static string Decrypt(this SecureString securePassword)
+        {
+            if (securePassword == null)
+            {
+                return string.Empty;
+            }
+
+            IntPtr unmanagedString = IntPtr.Zero;
+            try
+            {
+                unmanagedString = Marshal.SecureStringToGlobalAllocUnicode(securePassword);
+                return Marshal.PtrToStringUni(unmanagedString);
+            }
+            finally
+            {
+                Marshal.ZeroFreeGlobalAllocUnicode(unmanagedString);
+            }
+        }
+    }
+}