+++ SprintPlanner.WpfApp/Infrastructure/IHavePassword.cs
@@ -1,0 +1,9 @@
+﻿using System.Security;
+
+namespace SprintPlanner.WpfApp.Infrastructure
+{
+    public interface IHavePassword
+    {
+        SecureString Password { get; }
+    }
+}