+++ SprintPlanner.WpfApp/UI/Login/LoginView.xaml.cs
@@ -1,6 +1,5 @@
-﻿using MahApps.Metro.Controls;
-using SprintPlanner.WpfApp.Properties;
-using System.Windows;
+﻿using SprintPlanner.WpfApp.Infrastructure;
+using System.Security;
 using System.Windows.Controls;
 
 namespace SprintPlanner.WpfApp.UI.Login
@@ -8,11 +7,13 @@ namespace SprintPlanner.WpfApp.UI.Login
     /// <summary>
     /// Interaction logic for LoginWindow.xaml
     /// </summary>
-    public partial class LoginView : UserControl
+    public partial class LoginView : UserControl, IHavePassword
     {
         public LoginView()
         {
             InitializeComponent();
         }
+
+        public SecureString Password => userPassword.SecurePassword;
     }
 }