+++ SprintPlanner.WpfApp/UI/Login/LoginViewModel.cs
@@ -2,9 +2,14 @@
 using GalaSoft.MvvmLight.Command;
 using MahApps.Metro.Controls;
 using MahApps.Metro.Controls.Dialogs;
+using SprintPlanner.Core.Extensions;
 using SprintPlanner.Core.Logic;
+using SprintPlanner.WpfApp.Infrastructure;
 using SprintPlanner.WpfApp.Properties;
 using System;
+using System.IO;
+using System.Runtime.Serialization.Formatters.Binary;
+using System.Security;
 using System.Windows.Input;
 
 namespace SprintPlanner.WpfApp.UI.Login
@@ -13,6 +18,12 @@ namespace SprintPlanner.WpfApp.UI.Login
     {
         private readonly MetroWindow _window;
 
+        private ICommand _loginCommand;
+
+        private bool _storeCredentials;
+
+        private string _userName;
+
         public LoginViewModel(MetroWindow w)
         {
             _window = w;
@@ -20,31 +31,13 @@ namespace SprintPlanner.WpfApp.UI.Login
 
         public event Action LoginSucceeded;
 
-        private string _userName;
-
-        public string UserName
+        public ICommand LoginCommand
         {
-            get { return _userName; }
-            set
+            get
             {
-                _userName = value;
-                RaisePropertyChanged();
+                return _loginCommand ?? (_loginCommand = new RelayCommand<object>((parameter) => LoginCommandExecute(parameter)));
             }
         }
-
-        private string _password;
-
-        public string Password
-        {
-            get { return _password; }
-            set
-            {
-                _password = value;
-                RaisePropertyChanged();
-            }
-        }
-
-        private bool _storeCredentials;
 
         public bool StoreCredentials
         {
@@ -56,16 +49,15 @@ namespace SprintPlanner.WpfApp.UI.Login
             }
         }
 
-        private ICommand _loginCommand;
-
-        public ICommand LoginCommand
+        public string UserName
         {
-            get
+            get { return _userName; }
+            set
             {
-                return _loginCommand ?? (_loginCommand = new RelayCommand(LoginCommandExecute));
+                _userName = value;
+                RaisePropertyChanged();
             }
         }
-
 
         #region IsLoggedIn Property
 
@@ -87,30 +79,8 @@ namespace SprintPlanner.WpfApp.UI.Login
 
         #endregion IsLoggedIn Property
 
-
-        private void LoginCommandExecute()
+        public void Flush()
         {
-            Settings.Default.StoreCredentials = StoreCredentials;
-
-            IsLoggedIn = Business.Jira.Login(UserName, Password);
-            if (IsLoggedIn)
-            {
-                // Always store credentials, as a workaround to transmit them to main window
-                Settings.Default.User = UserName;
-                Settings.Default.Pass = Password;
-                Settings.Default.Save();
-                OnLoginSucceeded();
-            }
-            else
-            {
-                _window.ShowMessageAsync("Sprint Organizer", "Username or password is wrong.");
-            }
-
-        }
-
-        protected void OnLoginSucceeded()
-        {
-            LoginSucceeded?.Invoke();
         }
 
         public void Pull()
@@ -119,15 +89,44 @@ namespace SprintPlanner.WpfApp.UI.Login
             if (StoreCredentials)
             {
                 UserName = Settings.Default.User;
-                Password = Settings.Default.Pass;
             }
         }
 
-        public void Flush()
+        protected void OnLoginSucceeded()
         {
-
+            LoginSucceeded?.Invoke();
         }
 
+        private void LoginCommandExecute(object parameter)
+        {
+            Settings.Default.StoreCredentials = StoreCredentials;
 
+            if (parameter is IHavePassword passwordContainer)
+            {
+                IsLoggedIn = Business.Jira.Login(UserName, passwordContainer.Password);
+                if (IsLoggedIn)
+                {
+                    // Always store credentials, as a workaround to transmit them to main window
+                    Settings.Default.User = UserName;
+                    Settings.Default.Pass = Base64Encode(passwordContainer.Password.Decrypt());
+                    Settings.Default.Save();
+                    OnLoginSucceeded();
+                }
+                else
+                {
+                    _window.ShowMessageAsync("Sprint Organizer", "Username or password is wrong.");
+                }
+            }
+            else
+            {
+                _window.ShowMessageAsync("Sprint Organizer", "Unable to login. Error while reading password.");
+            }
+        }
+
+        public static string Base64Encode(string plainText)
+        {
+            var plainTextBytes = System.Text.Encoding.UTF8.GetBytes(plainText);
+            return System.Convert.ToBase64String(plainTextBytes);
+        }
     }
 }